flask/requirements.txt
Flask==2.3.3
google-cloud-secret-manager
google-generativeai
flask-cors